Add restore list
