+++ web/modules/custom/shepherd/shp_backup/shp_backup.links.task.yml
@@ -4,3 +4,9 @@ shp_backup.site_backups:
   route_name: shp_backup.backups
   base_route: entity.node.canonical
   weight: 10
+shp_backup.site_restores:
+  title: Restores
+  parent_id: entity.node.canonical
+  route_name: shp_backup.restores
+  base_route: entity.node.canonical
+  weight: 11
