+++ web/modules/custom/shepherd/shp_backup/shp_backup.routing.yml
@@ -46,6 +46,20 @@ shp_backup.backups:
     _permission: 'shp access environment wizard'
     _custom_access: '\Drupal\shp_custom\Controller\SiteLocalTaskController::checkAccess'
     site: ^[0-9]+
+shp_backup.restores:
+  path: '/site/{node}/restores'
+  defaults:
+    _controller: '\Drupal\shp_backup\Controller\RestoreList::list'
+    _title: 'Restore list'
+  options:
+    parameters:
+      site:
+        type: entity:node
+    no_cache: TRUE
+  requirements:
+    _permission: 'shp access environment wizard'
+    _custom_access: '\Drupal\shp_custom\Controller\SiteLocalTaskController::checkAccess'
+    site: ^[0-9]+
 shp_backup.settings:
   path: '/admin/config/shepherd/backup'
   defaults:
